+++ activerecord/lib/active_record/schema_migration.rb
@@ -45,7 +45,7 @@ module ActiveRecord
       end
 
       def all_versions
-        order(:version).pluck(:version).map(&:to_i)
+        order(:version).pluck(:version)
       end
     end